tello-client/requirements.txt
python-socketio[client]>=5.11.0
djitellopy>=2.5.0
opencv-python>=4.9.0
numpy>=1.26.0
